+++ src/pytypehint/__init__.py
@@ -14,7 +14,7 @@ from pytypehint.shapes import (
 )
 from pytypehint.utils import MISSING
 
-__version__ = "0.0.3"
+__version__ = "0.0.4"
 
 __all__ = [
     "struct_of",